requirements.txt
streamlit
groq
PyPDF2
exceptions
python-docx
python-dotenv